chore: ensure product quantity can be decreted

+++ src/actions/decrease-cart-product-quantity/index.ts
@@ -8,12 +8,12 @@ import { db } from "@/db";
 import { cartItemTable } from "@/db/schema";
 import { auth } from "@/lib/auth";
 
-import { removeProductFromCartSchema } from "./schema";
+import { decreaseCartProductQuantitySchema } from "./schema";
 
-export const removeProductFromCart = async (
-    data: z.infer<typeof removeProductFromCartSchema>,
+export const decreaseCartProductQuantity = async (
+    data: z.infer<typeof decreaseCartProductQuantitySchema>,
 ) => {
-    removeProductFromCartSchema.safeParse(data);
+    decreaseCartProductQuantitySchema.safeParse(data);
     const session= await auth.api.getSession({
         headers: await headers(),
     });
@@ -23,7 +23,7 @@ export const removeProductFromCart = async (
     }
 
     const cartItem = await db.query.cartItemTable.findFirst({
-        where: (cartItem, { eq }) => eq(cartItem.id, data.cartItem.Id),
+        where: (cartItem, { eq }) => eq(cartItem.id, data.cartItemId),
         with: {
             cart: true,
         },
@@ -38,5 +38,13 @@ export const removeProductFromCart = async (
         throw new Error("Unauthorized");
     }
 
-    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+    if (cartItem.quantity !== 1) {
+        await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+        return;
+    }
+
+    await db
+    .update(cartItemTable)
+    .set({ quantity: cartItem.quantity - 1})
+    .where(eq(cartItemTable.id, cartItem.id));
 };
